code comments, everywhere; correcting issues with cavern rotation timing reports

+++ src/main/scala/net/psforever/util/Config.scala
@@ -199,7 +199,7 @@ case class WarpGateConfig(
 )
 
 case class CavernRotationConfig(
-    hoursBetweenRotation: Int,
+    hoursBetweenRotation: Float,
     simultaneousUnlockedZones: Int,
     enhancedRotationOrder: Seq[Int],
     forceRotationImmediately: Boolean
